Add rebmark automation
Rationale: mkbmark is an automation, that abstracts the common case of
creating and maintaining bookmark aliases for a variety of utilies.

Well, rebmark is an automation of the common case of using mkbmark. It
allows the user to quickly recompile the bookmarks, so adjustments can
automatically be made in a live-system.