Add s automation for search utility
`s` is an alias to use search with $PWD as the first globbing pattern.
This restricts the search to the current directory. For use from the
terminal this is often more useful, than a global search.

For use as a global searcher, use `search` directly.

Note, that this automation will have to change, when the search impl is
extended to support multiple databases.